refactored hill climber tests
